Call BigDecimal(num) instead of BigDecimal.new(num) (#928)
Call BigDecimal(num) instead of BigDecimal.new(num) to be ready for
this deprecation in Ruby 2.6, and eliminate a deprecation warning now.

Reduce the number of extern classes - while developing this change,
I was chasing a NULL pointer on cBigDecimal from statement.c
because I removed the class lookup from result.c, but extern meant the
compiler did not flag it as an uninitialized value.